requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Vector Store
faiss-cpu>=1.8.0

# API Server
fastapi>=0.115.0
uvicorn>=0.30.0

# Env management
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0

# Testing
pytest>=8.0.0

# Embedding
sentence-transformers>=2.2.0

# Parquet file reading
pandas>=2.0.0
pyarrow>=12.0.0
langchain-huggingface>=0.1.0

